pridany test

+++ src/main/java/com/mycompany/calcularot/Calculator.java
@@ -2,7 +2,7 @@ package com.mycompany.calcularot;
 
 public class Calculator {
 
-    public int scitaj(int a, int b) {
+    public long scitaj(int a, int b) {
         return a + b;
     }
 }
